# [WIP] Use loaded custom_attributes in MiqReport#generate