@NickLaMuro 2018-06-20
Removing references to https://bugzilla.redhat.com/show_bug.cgi?id=1592480 since we are going with https://github.com/ManageIQ/manageiq/pull/17615 instead.  I will work on making this WIP as time permits.