@lpichler 2018-11-20
@NickLaMuro, it is instructive !

LGTM,

and is it alternative or improvement to your https://github.com/ManageIQ/manageiq/pull/17615 ?

thanks 👍 👍